Externalize CLI
This is just an intermediate, it’s about to all go in the main package
anyway
